pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-graphql-platform-processor"
version = "0.1.0"
description = "Azure GraphQL Platform Processor"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "DataLab"}
]
dependencies = [
    "azure-storage-blob>=12.25.1",
    "fastapi>=0.115.12",
    "pandas>=2.2.3",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "strawberry-graphql>=0.270.5",
    "uvicorn>=0.29.0",
    "polars>=1.33.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
azure-graphql-server = "main:main"

[tool.setuptools]
packages = ["gql"]
py-modules = ["main"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
